modified GenericObjectType to return None when attribute not found
